fix test operation.test.js
