+++ src/operation/operation.test.js
@@ -1,0 +1,34 @@
+import MockAdapter from 'axios-mock-adapter';
+
+import {ActionType} from '../constants';
+import api from '../api';
+import Operation from '../operation/operation';
+
+const mockApi = jest.genMockFromModule(`../api`);
+
+mockApi.get = jest.fn(() => {
+  return Promise.resolve({
+    data: [{fake: true}]
+  });
+});
+
+describe(`Reducer load data work correctly`, () => {
+  it(`Should make a correct API call to /hotel`, () => {
+    const apiMock = new MockAdapter(api);
+    const dispatch = jest.fn();
+    const offerLoader = Operation.loadOffers();
+
+    apiMock
+      .onGet(`/hotels`)
+      .reply(200, [{fake: true}]);
+
+    return offerLoader(dispatch, null, mockApi)
+      .then(() => {
+        expect(dispatch).toHaveBeenCalledTimes(1);
+        expect(dispatch).toHaveBeenNthCalledWith(1, {
+          type: ActionType.LOAD_OFFERS,
+          payload: [{fake: true}]
+        });
+      });
+  });
+});
