+++ src/reducer/reducer.test.js
@@ -1,17 +1,5 @@
-import MockAdapter from 'axios-mock-adapter';
-
-import {DEFAULT_OFFERS, SortNames, ActionType} from '../constants';
+import {DEFAULT_OFFERS, SortNames} from '../constants';
 import {getCityOffers, sortOfferList} from '../utils';
-import api from '../api';
-import Operation from '../operation/operation';
-
-const mockApi = jest.genMockFromModule(`../api`);
-
-mockApi.get = jest.fn(() => {
-  return Promise.resolve({
-    data: [{fake: true}]
-  });
-});
 
 describe(`Business logic is correct`, () => {
   it(`Get offers is correctly`, () => {
@@ -62,24 +50,3 @@ describe(`Business logic is correct`, () => {
   }]);
 });
 
-
-describe(`Reducer load data work correctly`, () => {
-  it(`Should make a correct API call to /hotel`, () => {
-    const apiMock = new MockAdapter(api);
-    const dispatch = jest.fn();
-    const offerLoader = Operation.loadOffers();
-
-    apiMock
-      .onGet(`/hotels`)
-      .reply(200, [{fake: true}]);
-
-    return offerLoader(dispatch, jest.fn(), mockApi)
-      .then(() => {
-        expect(dispatch).toHaveBeenCalledTimes(1);
-        expect(dispatch).toHaveBeenNthCalledWith(1, {
-          type: ActionType.LOAD_OFFERS,
-          payload: [{fake: true}]
-        });
-      });
-  });
-});
